# src/dao/models/Product.js
const mongoose = require('mongoose');

const productSchema = new mongoose.Schema({
    title: { type: String, required: true },
    description: { type: String, required: true },
    price: { type: Number, required: true },
    thumbnail: { type: String, required: false },  // Cambiado a false
    code: { type: String, required: true, unique: true },
    stock: { type: Number, required: true },
    category: { type: String, required: true },
    status: { type: Boolean, default: true },
    owner: { type: String, default: 'admin' } // Añadido owner
});

const Product = mongoose.model('Product', productSchema);

module.exports = Product;